requirements.txt
alembic==1.10.3
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.95.1
flake8==6.0.0
flower==1.2.0
greenlet==2.0.2
h11==0.14.0
humanize==4.6.0
idna==3.4
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
vine==5.0.0
wcwidth==0.2.6
